requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
python-dotenv==1.0.0
pymongo==4.5.0
pdfplumber==0.10.0
google-generativeai==0.2.0
certifi==2023.7.22
bcrypt==4.0.1